Covid 19 cases exploration
This project tells the total number of cases and deaths all around the world and also shows which country has the highest number of cases and deaths. And also how Covid 19 has affected both genders (males and females) and shows which country has the highest and lowest ratio of males: females cases.